fix issue for saving decimal values

+++ PalladiumProject.DTOs/DTOs/CompanySettings/LeaveSettings/LeaveSettingsResponseDTO.cs
@@ -13,7 +13,7 @@ public class LeaveSettingsResponseDTO
     public decimal? LeaveAccumulationDays { get; set; }
     public int? ExceedDue { get; set; }
     public bool? LeaveCarriedForward { get; set; }
-    public int? LeaveCarriedForwardMaxDays { get; set; }
+    public decimal? LeaveCarriedForwardMaxDays { get; set; }
     public bool? Recurring { get; set; }
     public int? NoOfTimeReccuring { get; set; }
 }
